Introduced function for reducing item data to keys and values.
The new function essentially simplifies the data for use with a
Mustache or Handlebars templates, getting rid of anything
extraneous.

I've kept the raw JSON output for each item in, for reference, but
these are now written to a folder called `raw`, whereas `items` is
now the reduced data form.