Simplified arguments for `getJSONfromAPI` function.
There used to be three arguments -- `request` for the path of the
API request, `output` for a filename to print the output to, and
`callback` for an asynchronous function call. I got rid of `output`
in order to separate GatherContent specific handling code from
more generic code like writing to files. Writing to the file can
be handled as part of any callback so this was a redundancy.

To prove this I have added the handling of raw JSON storage from
the API (which used to be handled by the `output` argument) into
the `processItem` callback which already contained the other
write to file calls.