Moved mocking of server to a set-up function.
All the Sinon-related code is now handled before each Mocha test
is run. This mocks the GatherContent API in a very basic way at the
moment and not accurately -- authentication is checked and if it
passes then the `options` argument stringified to the callback,
rather than a mock-data set.

Paths are not checked yet, although if authentication fails
the message "Invalid credentials." is passed to the callback
which does reflect GatherContent's current behaviour.