Added handling of data to sub-item retrieval in branch selector.
Now as sub-items are found and fetched, the data is stored in an
object within the larger branch selector closure.

The question remains as to how to then return this to the original
branch selector callback argument so that the user of the interface
can do what they need to with the data. Because all calls are
asynchronous, it is not enough simply to use the callback on
an incomplete item.