Added `id` fields to the sample data for use in `test/api-branch`.
This was not included before so was making it impossible for all
the tests to pass. Including it simulates what GatherContent also
returns when its API is called.

Also edited one of the unit tests to expect correct behaviour --
namely, even a leaf node has an `items` array albeit an empty one.