Added additional function called `flattenItemData`.
Behaves much the same as `reduceItemToKVPairs` but it expects
the actual `data` part of an item to be fed to it as an argument
wherease `reduceItemToKVPairs` will extract that part for a user.

There are no tests written for this new function yet -- I'm adding
it to see if I can resolve a problem around a project branch
unit test -- but in theory most of the inside of the
`reduceItemToKVPairs` should be rewritten with this new
`flattenItemData` function and all existing and already passing
tests should still pass.