Fixed test error in `getProjectBranch` where no item specified.
The solution was a result of abstracting item storage away from
navigating through various sub-/child-items. It meant that where
no item id has specified (and therefore defaults to zero), the
storage element can be bypassed, thereby ensuring that root remains
in tact and pointing to the right object through the trawling
of the entire project.