Removed error from `getProjectBranch` function to pass tests.
Now all tests should pass. The assignment of the results
of the by-item callback was going to `item.data`, a legacy from
an attempt to solve an old problem which has now gone. Having
this assignment in place was created a circularity within the
results (essentially making an item a property of its own `data`
object property) so getting rid of it made tests pass.