Added provisional solution for ordering branch items by position.
This should help pass some of the recently added unit tests but
does not deal with all cases. Some work needs doing around the
`position` property which is volatile owing to the nature of the
iterator function. Need to address that and help make its work
less destructive, even if it is provided by the user.