Renamed inner function and array in `getProjectBranch`.
Using `item_store` as the array name for storing new items in a
branch tier was getting confusing; I have renamed `sibling_store`
so that it's clearer when pushing an item to the array, any other
items in it will share the same parent.

Likewise the function name `getChildItems` was very confusing
as the function only deals with one item at a time and overall it
is not looking for a child in itself (though within it looks for
children of the single item specified and thereby calls itself
as many times as is needed). I simplified this to `getItem` which
hopefully makes the function's purpose clearer.