Changed unit tests to expect different behaviour.
Each leaf item is now expected to have an items property of its
own but it will be empty. This seemed neater than having nodes
without an items property -- it means each object has the same
properties potentially -- and it is easy to work out if the item
is a node or not simply by checking whether or not the items
property actually contains anything.